move JsonHelper to dedicated file
