+++ webapi/src/test/scala/org/knora/webapi/slice/ontology/api/service/RestCardinalityServiceLiveSpec.scala
@@ -2,8 +2,6 @@ package org.knora.webapi.slice.ontology.api.service
 
 import zio.test._
 import zio._
-import zio.json.ast.Json
-import zio.json.DecoderOps
 
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.messages.StringFormatter
@@ -23,25 +21,11 @@ import org.knora.webapi.slice.ontology.repo.service.OntologyRepoLive
 import org.knora.webapi.slice.resourceinfo.domain.InternalIri
 import org.knora.webapi.slice.resourceinfo.domain.IriConverter
 import org.knora.webapi.slice.resourceinfo.domain.IriTestConstants
-import org.knora.webapi.ApiV2Complex
-import org.knora.webapi.messages.util.rdf.JsonLD
-import org.knora.webapi.messages.v2.responder.KnoraJsonLDResponseV2
-import org.knora.webapi.slice.ontology.api.service.JsonHelper.renderResponseJson
-import org.knora.webapi.slice.ontology.api.service.JsonHelper.StringToJson
-import org.knora.webapi.slice.ontology.api.service.JsonHelper.parseJson
+import org.knora.webapi.util.JsonHelper.renderResponseJson
+import org.knora.webapi.util.JsonHelper.StringToJson
+import org.knora.webapi.util.JsonHelper.parseJson
 import org.knora.webapi.IRI
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
-
-object JsonHelper {
-  def parseJson(jsonString: String): Json = jsonString.fromJson[Json].getOrElse(throw new IllegalStateException)
-
-  def renderResponseJson(response: KnoraJsonLDResponseV2): URIO[AppConfig, Json] =
-    ZIO.service[AppConfig].map(appConfig => parseJson(response.format(JsonLD, ApiV2Complex, Set.empty, appConfig)))
-
-  implicit class StringToJson(jsonStr: String) {
-    def asJson: Json = parseJson(jsonStr.stripMargin)
-  }
-}
 
 object RestCardinalityServiceLiveSpec extends ZIOSpecDefault {
   private val ontology = OntologyCacheDataBuilder.builder
